last update

+++ src/Main.java
@@ -17,7 +17,7 @@ public class Main {
         c.write(data);
 
         System.out.println("Waiting for result...");
-        System.out.println("Result: " + c.readLong());
+        System.out.println("Result: " + c.readInt());
         curtask.end();
     }
